thread: fix destroying child processes
list_foreach (and even list_foreach_safe) are not resilient to moving a
list_node to anther list, as I need to do for moving process children to
init.

This change introduces a new list_concat primitive that moves a while
list at once by twiddling head and tail pointers. It's not super well
validated, but it does solve the problem at hand.