syscalls: don't cast all the function pointers to uintptr_t
This change removes the cast to uintptr_t in the syscall table. Function
pointers are implicitly convertible from
pointer-to-function-with-prototype to pointer-to-function-without, a la

void a(int a) {}
void (*fp)() = a;

I believe this is defined and legal, and it is accepted by all major
compilers with one exception -- very specifically GCC and Clang reject
this implicit conversion when the parameter list contains a parameter of
type _Bool. I cannot explain this, and expect it to be a bug in their
semantic analysis of the situation, but I have had to change the
prototype of sys_strace to take an integer rather than a boolean to work
around this bug. I have not changed userspace, and the ABI does not
change.