autogenerate syscalls
This is a pretty expansive change because it integrates the previous
commit's syscall autogenerator.

This meant I had to wrestle with places where I was inconsistent or
wrong, like having clone and time have different names in different
places.

The build story isn't fully developed, I'd like it to only run the
generator when SYSCALLS changes, but for the moment this is plenty fast
enough, and I'll fix it someday. Maybe when I have a build system I
don't hate. A boy can dream, right?