arch: goodbye x86_32
I find that x86_32 is just a burdeon on this project. It's nice to have
the seperation and check on introducing architecture-specific code, but
honestly it just stops me from wanting to make changes that touch
architecture code, since I can do all the work to get it working on 64
bit and then have to go again for 32 bit.

This should help unblock vmm upgrades, pmm upgrades, linker changes,
etc.