fs redo: it compiles
This is a major rework of how the filesystem works. Instead of files
having a name and a parent, these qualities are attached to directory
entries.

This inspiration for this was to have a directory child called "." and
".." for the magic directories that actually just point to the right
things.