libc: add atexit and experimental redirect output
atexit is pretty standard stuff
redirect_output_to is designed to allow a program to pipe its output to
another of its own accord. For example, I would like `ls` to colimate
its output, but I do not want to copy the colimation logic from `column`
to `ls` -- so, I have the binary send its own output to `column` when it
wants colimation.