fs/procfs: bring back real procfiles
This brings back real procfiles! Yay!

It's basically the exact same thing we had before, but with the fun
proc_sprintf function that makes it much nicer.

I still need to do the process stuff, becasue I do want those to be
dynamically generated and not managed manually like the "real" procfiles
are.

This also swaps the order of headers to make <basic.h> come first in
clang-format, so it touches a lot of other files too.