linker: replace with linker v2
This commit rips out the old hack linker and replaces it with the last
several months' work from the `linker` repository.

This version is much better organized in my opinion, it has a lot less
indentation and the metadata is better factored.

Everything seems to still work (loading modules and calling their
inits), except that I'm rethinking how pretty debug works, it won't be a
sprintf in the linker anymore, so that's not back just yet. `dump.rb`
understands this format too, so it's a noncritical setback.