pmm: add something like OOM handling
This kills the process that allocates the last available page, and adds
a test program that tries to fill memory to test it. The program runs
and dies, and then there is available memory, so it works at least
somewhat.

This is a regression in the case of something like 'bomb', because it
will continue making new processes that each try to touch the last page
of memory until something bad happens.

This also should be improved to handle the case where the last page of
memory is allocated by the kernel itself, but I'll leave that for the
future.