ls: add new filetype sigil for PROC_THREAD
This was a debugging adventure. The bug was that `ls` would magically
miss any directory children after the first PROC_THREAD, but only if it
was printing sigils *and* outputting its output to `column`, *and* not
if you piped it.

    $ ls -al /proc/1 | column
    .    ..   comm pid
    $ ls -a /proc/1
    ./   ../  comm
    $

This was happening because when you use ls without -l, it uses the
interactive mode but sends stdout to column automatically. This stdout
is buffered, and column relies on `fgets` to pull strings from the pipe
one line at a time.

The fgets code (specifically consume_until) has a bug where it checks
for the delimeter (and existance of a delimeter) with strchr, which
stops at the first NUL (as with all C string functions), so if there's a
NUL in the buffer it will never find a delimeter after the NUL
character. This meant that column read all remaining entries at once
when it hit EOF and the buffer was returned anyway, but the NUL foiled
us again by preventing anything after it from printing, making it look
like we only read `comm` and nothing else.

Why was there a NUL character in the buffer? I spent a long time looking
at the fstdio code trying to find an off-by-one or something, because
the `readdir` code is definitely just formatting a NUL-terminated
string into the ng_dirent buffer like everything else, and I did find
bugs (which are included in this CR), but nothing to cause the observed
behavior. Furthermore, this didn't happen when manually piping to
column, so it couldn't be a foundational problem reading the data
structures.

Rather, it turns out the problem was that I was trying to add a filetype
sigil for a filetype that didn't exist in `ls.c` yet, which was reading
off the end of the array and happening to find a -- '\0'.

The formatted file names going into the pipe were thus:
    "comm\0\npid\0\n"
which triggered the fgets bug and stopped the stream processing any
further.

Resolution: add a sigil for FT_PROC_THREAD
Future work: figure out how to not do this again