fs: use real directories in initfs, move everything to /usr
Instead of delivering a flat tar with a bunch of stuff that we yolo into
/bin, this changes it to tar up the entire sysroot as it should be.

I do ignore directories that aren't created manually in vfs_init, but
this does get us distinct usr/bin, usr/lib, and usr/include directories.

It's all under /usr mostly becuase I'd have to recompile GCC to change
the sysroot if I changed it, so I'm going to be okay with this for the
moment, though it's open to be reconsidered at some point in the future.