exec: consolidate bootstrap_usermode as an exec
This has some interesting performance implications, it seems like
starting usermode is super slow now - but I don't believe it -- it looks
to me like this has just moved the expensive parts around so they're
more visible, but the time appears to be inside
"vmm_create_unbacked_range", which I suppose makes some amount of sense,
but also is pretty dissapointing that we're hitting noticeable slowdowns
there.

This is quite apparent on startup and should convince me to actually
work on the performace considerations, so I'm going to go ahead and
commit this.