everything: fix random triple faults
I was getting random triple faults, and it seems to have been caused by
me switching the timer over to use the system heap. There is either some
bad dependancy problem between sync and the heap, or possibly just me
hammering it too much.

This change switches it back over to spalloc, having been resurected
from the dead.

I do need to figure out what's wrong with the allocator / the system
when using that as the backend for timer_events. Having the system start
randomly triple faulting after a lot of major refactors is a really bad
time, and I should find some way to address that.

I know linux does the thing where you have a totally seperate stack for
double faults, so the thing at least doesn't triple fault -- that makes
it super hard to debug since the debugger is just gone after the VM
reboots. I'll need to think about supporting that.

It's not a good day when you open the Intel manual, but that was this
one. Here's hoping it doesn't happen again soon, and/or that I actually
fixed it this time.