procfs: remove half-baked "thread_file"
This replaces the not-quite-working thread directories and inner files
with a thread procfile for everyone.

I can just make this a collimated table of the necessary information and
then actually use it in places like `top`, rather than waiting on
inspiration for a linux-style procfs.

Whether I do that later or not remains to be seen, but this works for
now.