socket: clean up some socket_close behavior
This is a first attempt at fixing closing sockets not properly waking
people who are waiting on others to write to those sockets. It's not
completely working yet, mostly I think because there's a fundamental
difference between sockets and something like pipes where it's very
clear when you need to wake up all the readers because the writers have
gone away. Since sockets are both multiple file descriptions to
represent a single notional connection and also both sides have RW, it's
more complicated to track the state and this doesnt do it correctly yet.