sh: release the experiemental shell
This is a major improvement to the shell.

Parsing is much much much better, the tokenizer is much much better,
everything is just much better organized and put together.

Basically for free, this upgrade also brings along some fun new
features, including habing multiple commands on a single line seperated
by ';', and the logical operators && and ||, with their correct
short-circuiting behavior.

I intend to continue to expand this shell, it's much more approachable
than the old implementation was, and should prove much more amenable to
extension.