x86/vmm: don't stop the world during fork
I'm not confident this is okay, but it seems to work.

This was originally here because the fork map was a global, and couldn't
be stopmed on by another fork, but that's not true anymore.

I've been running this for a while and it doesn't seem to cause huge
problems, but I suppose we'll find out in time.

The main worry is that if there are threads modifying the memory map at
the same time as a fork is done, that we'd see a torn version of memory,
but I think that's ineviatble. If you fork while another thread is
writing something, you will certainly see a torn version of that thing,
dependant on when the precise COW cutover happened. -- I believe this is
why the fork(2) manual page specifies that POSIX requires child
processes forked from parents with multiple threads to only perform
sync-safe actions before calling execve(2).