rust/sync: remove the half-baked notion of thread-waiting
This just rips out any thread-waits because they did not interact well
with the concurrency primitives already in the C code in the kernel. I
intend to get some semblance of a `join` back by using a Condvar and
implementing that entirely in Rust code.

That's a job for the future.