event_log: don't use mutex for the moment
The mutex ends up being problematic because I currently log an event
when threads switch but if the lock is taken you can't get to another
thread if that lock is taken.

Likely the ultimate solution will be to put the event logging code in a
critical section, but we're not there just yet.