thread: don't run runnable checks in enqueue.
It turns out you can enqueue threads that aren't runnable -- consider
the case where a thread is asleep and C-c is sent. That thread needs to
be able to run `handle_pending_signals()` in `thread_switch()`, but
can't if it can't be enqueued. In fact, right now the OS crashes if a
signal is sent to any thread that isn't runnable.

This fixes that by reinstaating the `if (!runnable) block;` check in
thread_switch, but may reintroduce old bugs with threads getting dropped
if there's a race condition in the state-swapping code.