interrupt: print a backtrace of the user program when we can
Previously, we would stop a backtrace at the kernel/user boundary when
an interrupt happened inside the kernel as a result of a syscall or
other user/kernel interaction. This makes it so if we are in the kernel
but have a valid user context we continue the traceback through user
code. That traceback isn't pretty and doesn't show symbol names, but it
is enough to be able to run dump.rb and get information about what is
going wrong.