stdio: significant rework of FILE stream buffering
The code is better, more accurately matches the specification, actually
sends output through buffers, and is more understandable.

This took a few revisions, but it always helps to return to designing
the interface first. I was originally trying to rework buffer first with
its own struct, but since I had no idea what I needed it was
fundamentally unfit (seperate read and write buffers in the same stream
for example). I regrouped, wrote fread and fwrite first, and was able to
get a really nice design that way.