tty: correct handling of C-d
C-d is not supposed to blanket kill the stream, it's a flush if there is
data remaining in the buffer and only an EOF if there is not.

Honestly that probably means it's not a special case, since returning 0
from a read is an EOF, but I'm going to stick with the signal_eof
mechanism I already have since it seems to work.

Worth considering in the future.