thread: make the kernel non-reentrant
This is a pretty major overhaul. Since the beginning of time, I have
been assuming the system can suspend nearly anywhere, but I think as
time goes on this stance becomes more and more untenable.

I've been wrestling with a bunch of concurrency problems for years at
this point, and known about lots of places where I need to act and then
atomically take a lock, but that's not always possible.

This change builds on the previous work to replace critical-section
locks with spinlocks that aren't reentrant and allows the kernel to
(essentialy for the moment) opt-in to being scheduled. This is
accomplished principally by stopping IRQs on syscall entries to the
kernel. As is here, this means the only times interrupts are enabled
are:
1. when running usermode code
2. when running thread 0 (idle thread)

Thus, not counting wait conditions, no kernel code can be interrupted
while operating.