apic: use the right address for the local apic
I had the wrong address in the lapic address variable, which meant
nothing was working.

This change backs out some of the debugging we were using to diagnose
what was wrong as it is no longer needed.

This also re-adds vmm_resolve, which was removed in one of the refactor
passes.