libc: provide syscalls as __ng_(name) with a weak alias as (name)
This allows me to make syscalls with the same name as libc functions
without creating name collisions or having to resort to musl-style
"__syscallN(NG_name)" for all real syscalls.

If a syscall is defined in interface/, by default export that name as a
function callable from C, but if an alternate definition is provided in
libc that will override the default alias.

As a concrete example, if I want to add logging to read(2) in userspace,
I need only define a read(3), such as

    ssize_t read(int fd, void *buffer, size_t len) {
        printf("do some logging\n");
        ssize_t result = __ng_read(fd, buffer, len);
        if (result < 0)
            printf("it failed!\n");
        return result;
    }

This was super annoying to do before because the "read" name was a
strong symbol and there was no other. Combine that with the syscall
generation system, and I would have to rename the entire syscall to do
some simple logging.

This is somewhat fragile to usermode programs defining their own
overrides for syscalls, but I own the whole system, so I'm not too
worried about that. Without these being weak aliases that case would
have failed to link anyway, so I'm not breaking any legitimate cases for
duplicated function names, and in fact I'm allowing userspace to define
names they couldn't before as long as they don't need those specific
system calls.