stdlib: remove datastructures from headers and namespace better
I didn't like having mregion and mheap in the stdlib.h header, if we
really do need to expose those let's make a seperate header or something
for dedicated heaps. This works with everything I was already doing.

While I was at it, I'm going to move things outside the POSIX namespace
to reserved names in libc, so a few collideable names become __names and
some internal functions are made static. Should be a boon for the
optimizer too.