build: add an option to build using nightingale-gcc in cmake
This also fixes a number of issues that have cropped up in gcc-compat
while I wasn't using it. At the moment, gcc builds but does not run.

I wanted to do this to take advantage of some of the more advanced
analysis tooling being built on gcc recently, like -fanalyze and
-Wstrict-aliasing.