thread: remove most irq enable/disable dancing
These have been wrong since I made the kernel a cooperative multitasking
system, since the enable and disables happening in thread_switch were
not reflecting the CPU's disabling IRQs in interrupts. In fact, these
were probably wrong long before that for the same reason.

This change removes most of the juggling - in future if I want
preemption I'll do it opt-in - having specific places where threads
elect to be preempted instead of allowing it willy-nilly anywhere.

This should also match much more accurately when interrupts are
_actually_ enabled for usermode.