thread_local storage: proof-of-concept
This concept relies on running on an Ivy-Bridge or newer CPU. There are
probably other ways to set the FS segment base, but here I'm using
wrfsbase which was added by Intel with the Ivy-Bridge microarchitecture.

The test program thread_local creates single-integer thread-local
storage allocations in two threads, performs operations on each
seperately, and then examines the backing storage to show the system
working as designed.