fgetc: prevent early EOF
fgetc is supposed to return a positive character wieh there is real data
available and -1 (EOF) on error or EOF condition.

I was reading the respective character into a signed char, then allowing
it to promote to int, which in turn meant there was a 1/256 chance the
function would spuriously return EOF (for random input data).

This fixes that by changing the internal char to be unsigned, changing
the promotion rules to not sign-extend and thus leaving all valic
characters positive.