fstdio: always remove files from all_files list before freeing them
If an attempt was made to open a FILE and an error occurred during the
underlying open(2) call, the FILE would be freed but would still be a
member of the __all_files list. This would then cause the process to
crash later during the finalization process where all open files are
closed due to the invalid member of the __all_files list.

This centralizes FILE deletion on _delete_file, which is next to
_new_file and can mirror its functionality to prevent this going
forward.