thread: move child reparenting to do_process_exit
This fixes a longstanding bug where when child processes would exit,
they would attempt to send SIGCHLD to an already-freed parent. They did
this because the now-zombie parent would hold onto its children until it
was collected by waitpid(2), but now I reparent a process's children to
init as soon as all of its threads are dead.

This makes a lot more sense, since if a process has no running threads
it can't service a SIGCHLD anyway, so really this is the most sensible
option.