multicore: initialize second core to C code
This is a super naive implementation of a second core - not dynamic at
all, this just creates a second hardcoded copy of the GDT and TSS to run
the second core on.

The second core also can't do much of anything, since it has no timer
and no interrupts routed to it, but it can print a nice message and then
halt.