thread: use GS for a CPU-local structure, not thread-local
This converts GS from a pointer to running_thread to a pointer to the
new this_cpu. The reason for this is to give me a place to put CPU-local
variables in a way that code can find it.

For now, the main customer for this is just to store the running thread
in ->running, but the idea is that I can add more later as it's needed.