include: prevent most transitive includes of <basic.h>
This moves the most critical (and safest) parts of basic.h to the
semi-standard sys/cdefs.h, which allows me to not include basic in
(nearly) any headers.

This should help the namespace polution situtation quite a bit.

This makes a number of fixes to places where the transitive dependency
was required.