[boot] (Start) Replacing GRUB with Limine
Grub has the distinct disadvantage that it doesn't run on my machine,
and Limine worked almost immediately. I haven't transitioned all the
data previously gathered from multiboot, but this is a good start and
the operating system _does_ boot and gets to the multiboot magic number
check before crashing.

This does remove the skeleton I build to boot an applicaiton processor,
but that didn't work anyway, so it's not a huge loss.