c++: remove unnecessary extern "C"s
I didn't realize that if a declaration is made with extern "C" and then
the definition doesn't have it the object is in fact extern "C". That's
nifty because it means everything defined in the headers is
automatically extern "C" for C and everything that isn't is internal and
can use C++ features.