c++: call global constructors.
This sets up the .init_array array so I can call global constructors.
I'm currently choosing to do this right before jumping to the "real"
kernel, which seems like a reasonable place to do it, since we have
serial and malloc available.

There's probably room to valicate that systems have been brought online
in future so we can assert on anything being called from constructors
that isn't ready.

This was also doing... weird things with the thread.cpp initializer
function and causing problems later on, so I've made sure proc_zero
doesn't require a global initializer for now. We'll handle this next
time it comes up. Or maybe the problem will go away!