signal: define userspace as responsible for __ng_sigreturn
This represents a dramatic simplification of signal handling logic --
previously I was asserting that the kernel should make a sigreturn stub
available in every process memory space and set up the return for you,
so you could call sigaction on any function and get what you'd expect.

This commit changes that assumption -- now userspace is expected to
explicitly call sigreturn. You could do this in every signal handler,
but I've chosen to expand libc's signal handling to do this for you.